backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
langchain-groq==0.2.0
langgraph==0.2.28
langchain-core==0.3.10
python-multipart==0.0.9
aiosqlite==0.20.0
pymysql==1.0.3
